+++ src/controllers/purchase.controllers.js
@@ -1,8 +1,26 @@
 const catchError = require("../utils/catchError");
 const Purchase = require("../models/Purchase");
+const Category = require("../models/Category");
+const Product = require("../models/Product");
+const Cart = require("../models/Cart");
 
 const getAll = catchError(async (req, res) => {
-  const results = await Purchase.findAll();
+  const userId = req.user.id;
+  const results = await Purchase.findAll({
+    where: { userId },
+    include: [
+      {
+        model: Product,
+        attributes: ["title", "price"],
+        include: [
+          {
+            model: Category,
+            attributes: ["name"],
+          },
+        ],
+      },
+    ],
+  });
   return res.json(results);
 });
 
@@ -14,9 +32,13 @@ const create = catchError(async (req, res) => {
     raw: true,
     attributes: ["quantity", "userId", "productId"],
   });
+  if (!cart) return res.sendStatus(404);
 
   const result = await Purchase.bulkCreate(cart);
+  if (!result) return res.sendStatus(404);
+
   await Cart.destroy({ where: { userId } });
+
   return res.status(201).json(result);
 });
